+++ ThreeJS.novaextension/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Version 1.1.0
+
+- Added color picker support for Three.js colors in HEX and HSL format
+- Color swatches now appear next to Three.js color values
+- Click on colors to edit them with Nova's color picker
+
 ## Version 1.0.1
 
 - Initial release